// src/lib/moetruyen/model/getV1Manga200DataItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Moetruyen Public API
 * Unofficial read-only REST API for [MoeTruyen](https://moetruyen.net/).

Full features will be added in the future (chắc thế)

Github: [TNTKien/moetruyen-public-api](https://github.com/TNTKien/moetruyen-public-api), [dex593/web1 (MoeTruyen)](https://github.com/dex593/web1)

NOTE:
- All API endpoints have a global rate limit of 7 requests per second per IP.
- To avoid future issues, include the Origin: https://suicaodex.com or https://moetruyen.net headers when making API requests.
 * OpenAPI spec version: 0.1.0
 */
import type { GetV1Manga200DataItemGenresItem } from "./getV1Manga200DataItemGenresItem";
import type { GetV1Manga200DataItemStatus } from "./getV1Manga200DataItemStatus";

export type GetV1Manga200DataItem = {
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  id: number;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  description: string | null;
  author: string | null;
  status: GetV1Manga200DataItemStatus;
  cover: string | null;
  coverUrl: string | null;
  coverUpdatedAt: string | null;
  latestChapterNumber: number | null;
  latestChapterNumberText: string | null;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  chapterCount: number;
  isOneshot: boolean;
  genres: GetV1Manga200DataItemGenresItem[];
};